OTAT-464: Convert BigDecimal to double

+++ src/main/java/org/flickit/flickitassessmentcore/domain/AnswerOptionImpact.java
@@ -5,15 +5,14 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @EqualsAndHashCode
 public class AnswerOptionImpact {
+
     private Long id;
-    private BigDecimal value;
+    private double value;
     private Long optionId;
     private Long questionImpactId;
 }
